Fix SavedExtendedRegisters not handling init optimization
As it turns out, it is possible for an x86_64 CPU to legally fail to
save the state of a particular component if that component is in its
default state. SavedExtendedRegisters did not handle this, potentially
returning junk when reading such registers and failing to write them.

Proper handling for this has now been added. Reading a register of a
component that has not been saved will yield its default value and
writing such a register will properly initialize the relevant component
in the XSAVE block.