Fix incorrect use of raw spin::Mutex in Future
Previously, the implementation of futures used the raw Mutex type from
the spin crate to guard its shared internal data. This would prevent a
future from ever being fulfilled from an interrupt handler, since the
interrupted thread could be holding the lock on the internals of the
future, causing a deadlock.

Instead, an InterruptDisableSpinlock should be used, which disables all
interrupts on the current core while it is locked. This prevents any
such deadlock from occurring.