Add the ability to view any future as a Future<()>
Previously, the kernel's implementation of futures did not provide any
way to view multiple futures returning disparate types in a uniform
manner. This would make it impossible to ever create an API that did
something like wait for the first of a number of futures to be resolved.

To make APIs like that possible, it is necessary to be able to create a
Future<()> corresponding to any future. Effectively, this allows one to
create multiple futures that will resolve to () whenever the original
future is resolved, even if the type of value returned by the future
does not implement Clone.