Fix unsoundness in ProcessLock::create_kernel_thread
It turns out that the trait bound used by the create_kernel_thread
function was unsound, as it did not require the closure to have a static
lifetime nor did it require that the closure be Send. This would allow
closures to be passed that referenced non-Sync data or that referenced
data that may go out of scope while the thread is still alive.

To provide a simple escape hatch for avoiding the 'static bound (mostly
for the purposes of testing), a new create_kernel_thread_unchecked
function has been added that does not require that the closure be
'static.