Fix fsbase and gsbase being clobbered by interrupts
Previously, the interrupt handling routine would not properly save and
restore the fsbase and gsbase MSRs when switching into a kernel context,
which would result in them getting reset to 0 whenever an interrupt
occurred. This would manifest as page faults when trying to access
thread locals on kernel threads during or after an interrupt.