Fix VGA text buffer using lower-half virtual address
Previously, the VGA text buffer would be initialized with a virtual
address of 0xb8000. This was fine previously, as the bootloader leaves
this address mapped. However, this virtual address is in the lower half
of virtual memory, which is generally reserved for user-space mappings.
Instead, the mapped physical address in upper-half should be used.