Implement support for a custom GDT
Previously, the HydroxOS kernel would use the GDT set up by the
bootloader. While this works for simple cases, it has several drawbacks
that would make it a bad idea for the long-term.

Firstly, the GDT from the bootloader has a virtual address that is in
the lower half of memory, which should be reserved for user-mode
programs to make use of.

Secondly, the GDT from the bootloader does not provide any descriptors
besides a kernel CS and DS. For user-mode to be implemented, a user CS
and DS will be required.