Add assert that kernel is loaded in higher-half
Due to an issue with I was running cargo at one point, the linker script
that forces the kernel to be linked in higher-half was not used. This
resulted in the kernel being loaded in lower-half, which causes triple
faults once the lower-half mappings were removed from the page tables.

While this shouldn't happen in practice, init_kernel_addrspace has been
modified to ensure that the kernel is loaded in higher-half before
removing the lower-half mappings. This should at least make any such
future issues easier to debug, since they'll cause a panic instead of a
triple fault.