Allow frames used by the bootloader to be reused
Previously, frames marked in the region table with type Bootloader would
not be added to the frame allocator upon initialization. This was
necessary in the past, since we were still using the bootloader's GDT,
which resides in such memory. However, now that we use our own GDT,
bootloader frames should be safe to reuse.