Add an architecture API module
Previously, many parts of the kernel used the x86_64 package directly.
In order to reduce the number of architecture-specific parts of the
kernel, it is necessary to write these parts against an API that is
architecture-agnostic.

To facilitate this, a new module has been added that contains a dummy
architecture implementation with only the API surface, with all
functions unimplemented. By using cargo check with the check_arch_api
feature flag set, the architecture-agnostic parts of the kernel can be
checked to ensure they only use architecture-agnostic APIs.

Note that there are still a couple parts of the kernel outside the
arch::x86_64 module that are x86-64-specific. These are currently
guarded using check_arch_api, but the eventual goal would be to contain
all of this code in the architecture module.