Fix memory leak in Future::all
Previously, the Future::all function would increment the number of
references to the inner wait structure for each future passed in but
would never decrement it.

It turns out that the increment was unnecessary in the first place.
Keeping the reference alive is already handled by the FutureWriter being
consumed by the final future to resolve. Removing this increment allows
the newly created future's internal data to be immediately freed if all
of the futures passed in were already resolved.