Implement basic early de- and reallocation
Previously, the early allocator used a very naive bump pointer
allocation method. While this was fine before, the use of format! and
futures in logging means that a number of temporary object can get
allocated, reallocated, and deallocated in rapid succession.

As a minor optimization, the early allocator can now handle shrinking
allocations in place without needing to allocate new memory and can also
handle growing and freeing the last allocation made in-place.