Allow testing to continue after a failure
Previously, the kernel's unit tests would always abort after the first
failure. This was necessary since panics cannot unwind due to missing
library support and generally because doing so would be unwise in a
kernel.

However, it would be nice for the test suite to continue running when
possible after a failure, since most test case failures do not render
the system inoperable and stability is less of a concern when running
the test suite.

In order to do this, the behaviour of panics has been changed when
running the tests and the tests now execute on a separate thread that
will be killed if a test case fails. However, this only works if the
test thread was not holding any uninterruptible spinlocks at the time of
the panic. If the thread were holding one, state could be corrupted if
the test thread were aborted.