Fix Thread::suspend_current restoring IF in interrupt handler
Calling the Thread::suspend_current function will suspend the current
thread, releasing its spinlock in the context of the interrupt handler
that will perform the necessary context switch.

As an unintended consequence of this, the last InterruptDisabler was
being released in the context of this interrupt, which would result in
the previous state of the interrupt flag being restored during the
interrupt. This could result in interrupts unexpectedly being enabled
while an interrupt is already being handled and interrupts not being
re-enabled when the calling thread is resumed.