Fix incorrectly aligned TLS under certain conditions
As it turns out, the linker script used by the kernel did not correctly
ensure that the TLS's size had the minimum expected 16-byte alignment.
This could result in the size of the TLS being misaligned, which would
result in every variable in the TLS being misaligned, invoking undefined
behaviour.