Implement spinlock tracking
Previously, the kernel didn't perform any tracking of what spinlocks
were held by the current CPU core. This resulted in a number of hangs in
the kernel caused by attempting to lock a spinlock already locked on the
current CPU core.

In order to aid in debugging, a new feature has been added to allow
tracking of which spinlocks are currently held. If an attempt is made to
lock a spinlock which is already held by the current CPU core, the
kernel will panic instead of deadlocking.