Switch to using build.rs to specify the linker script
Previously, the linker script was being provided via the config.toml
file. While this worked, it had one major disadvantage: it would not
automatically rerun the linker if the linker script was changed. This
resulted in a lot of wasted time due to the resulting confusion.

To address this, a build.rs script has now been added that will specify
the linker script and inform cargo to rerun it when the linker script
changes. This effectively results in cargo properly rerunning the linker
when the linker script is changed.