Fix potential UB due to ThreadWaitList in FutureWaitGeneric
Previously, the implementation of futures in the kernel used a single
data structure called FutureWaitGeneric to store all state information
about a future, including the list of waiting threads.

However, as it turns out, having a ThreadWaitList inside of a data
structure protected by a spinlock and releasing the spinlock before
suspending the thread is actually undefined behaviour. This is the case
since the data inside a spinlock is given by &mut T when locked, meaning
that no other threads should ever access this data while it is locked.
ThreadWaitList violates this principle since the thread that's waiting
keeps a reference to this waitlist around inside its scheduler data
structure. While not currently used, it may be used in the future to
implement interruptible waiting.

Because of this, it's now necessary to place the ThreadWaitList outside
of the data structure guarded by the spinlock. Additionally, the
interface for ThreadWaitList has been updated to ensure that the
reference passed to wait() is valid until the thread is suspended.