Make test thread stack size a single const
Previously, a number of tests would create threads with stack sizes
defined as a literal number of bytes. This lead to difficult-to-debug
problems in some corner cases where the stack would overflow. In order
to make it easier to potentially raise the stack size for such threads
in the future, it has been put in a const called TEST_THREAD_STACK_SIZE.