Fix broken threads lists
Previously, a bug during thread creation would result in the
doubly-linked list of a process's threads getting corrupted. This could
result in a variety of very bad symptoms when these threads were being
deleted.

This bug has been fixed and a number of safety asserts have been added
during thread removal to ensure that the doubly-linked list of threads
is not corrupted in such a manner.