Abort testing if panic occurs in interrupt or another thread
Previously, the test framework would try to continue testing anyway if a
panic occurred during an interrupt handler or on a thread other than the
test thread. This was not desirable as it would almost always lead to
another panic, a page fault, or a deadlock.

Instead, special checks have now been added to detect these cases and
abort testing immediately if they are detected.