Add a very basic thread-based mutex
Previously, the only synchronization primitives available in the kernel
were based on spinlocks and disabling interrupts. This is fine and even
necessary for many critical data structures that will be locked for very
short periods of time, however it could result in a lot of useless
spinning in SMP systems (once I get around to implementing those) when
the lock is held for too long.

In order to avoid this useless spinning, a new Mutex synchronization
primitive has been added that makes use of the scheduler's capabilities
to suspend the current thread while waiting for the lock to be released.