Remove restriction on T for DeviceLock<T>
Previously, DeviceLock<T> had a restriction on the struct itself that T:
Device. This made it impossible to upcast a DeviceLock<T> to any trait
object that doesn't have Device as a superpredicate.

This was unnecessarily restrictive, since the Device bound is really
only required on the concrete type: many methods on DeviceLock<T> don't
need to call any Device methods. In light of that, the bound has been
loosened.