Implement DynDynBase for DeviceLock<dyn Device>
In preparation for a future version of dyn-dyn that will add basic
support for DST structs, DynDynBase is now implemented for
DeviceLock<dyn Device>. This should allow for downcasts from this type
to other implementing traits (also wrapped in a DeviceLock).

While dyn-dyn will be adding a way to safely derive DynDynBase for some
DST structs in the future as well, that will not work in this case. The
reason is that the support for that will only work if the type of the
last field also implements DynDynBase itself. That's not possible in
this case, since the last field is enclosed in a spinlock meaning that
calling methods on it directly isn't possible. An unsafe manual impl
with the DynDynTable stored elsewhere is the only solution to this.