Allow creating new Futures from a FutureWriter
Previously, the only way to create a Future corresponding to a
particular FutureWriter was to keep around a clone of the Future
returned when Future::new was originally called to create the
FutureWriter. This was somewhat inefficient, because it required that
the FutureWait data pointer be stored twice in some cases. Instead, a
new method has been added to FutureWriter that allows for creating a new
Future for its value after its original creation.