Switch to using shared refs for device methods
Previously, the Device trait and its subtraits would have methods
declared as &mut self. This was relying on the fact that the outer
container for the device was the location of the spinlock. However, this
was potentially problematic in some cases, because device methods may
want to drop their own locks to call methods that could potentially
deadlock otherwise.

In order to accommodate this, it is now the internal device's
responsibility to handle any locking that's necessary for that given
device to work properly. This provides more flexibility for the device
to handle special cases without risking deadlocks and more efficiently
in some cases.