Switch to using raw spinlocks in UninterruptibleSpinlock
Previously, UninterruptibleSpinlock<T> would internally use a
SpinMutex<T> to control access to its contents. While this was
convenient, it meant that it was difficult or impossible to implement
certain debugging functionality that would need to access spinlocks
without regard for what they contained. It also meant tying a lot of
functionality directly to the spin crate, which would make it difficult
to swap out later if need be.

In order to rectify this, a new RawSpinlock type has been added, which
uses a spin::Mutex<()> to perform locking. UninterruptibleSpinlock<T>
has been rewritten to use this instead. This means that it is now
possible to perform operations on all types of spinlocks uniformly by
simply performing operations on the RawSpinlock instead.