Implement mapping of spinlock guards
Previously, it was only possible to construct a spinlock guard referring
to the exact contents of a spinlock. However, for devices (which use
dynamic traits), it is sometimes necessary to return a reference to data
that is inside of a spinlock but where the exact shape of the internals
is unknown to the user.

To accommodate this, a new system has been added wherein an
UninterruptibleSpinlockGuard can be mapped to refer to a subset of the
data that it actually contains and made read-only. This allows such a
reference to be returned using a concrete type without any additional
allocations.