Work around inline asm issue with pushing gs/fs
For some reason, it seems that the handling of the instructions to push
and pop gs/fs in the inline assembly common interrupt routine has
changed in newer versions of LLVM. These now seem to encode as pushw and
popw instead of quadword pushes and pops, which results in the interrupt
frame being misaligned from what the Rust interrupt handler expects.
This was leading to crashes due to the interrupt number read not being a
valid one.