Add soft interrupts
Sometimes, it's necessary to delay performing a certain task until all
currently held spinlocks are released to avoid deadlocks. This is
especially true when dealing with futures, since when_resolved will be
called immediately upon FutureWriter::finish being called, which might
very well be done while critical spinlocks are held.

To allow this, a new feature called "soft interrupts" has been added to
the scheduler. This allows enqueueing a function to be called later,
either at the end of the current interrupt handler or when the last
InterruptDisabler is dropped (and so all uninterruptible spinlocks are
released).