Add UninterruptibleSpinlockGuard::replace_data
Sometimes, it's necessary to return an UninterruptibleSpinlockGuard to
keep a spinlock active while having it dereference to completely
unrelated data that's not intrinsically guarded by the spinlock. This
generally happens in unsafe code when some of the data logically guarded
by the spinlock isn't actually inside the spinlock itself (such as with
intrusive linked lists).

Previously, doing this using UninterruptibleSpinlockGuard::map required
the map itself to be unsafe due to some liftime constraints on that
method that ensure that the mutable reference passed in cannot be
leaked. However, there's nothing inherently unsafe about such an
operation as it doesn't actually leak the mutable reference to the
spinlock internals. To provide a safe API for this operation, we just
need to add a way to replace the data reference directly without using a
function called with a reference to the spinlock internals.