Clean up how testing is handled in kernel's main.rs
Previously, the implementation of kernel_main would be completely
replaced when running tests on the kernel binary. This was causing some
spurious warnings about dead code when running 'cargo check' with the
tests enabled, since the normal kernel_main was excluded in this
configuration. It was also making it more difficult than necessary to
work on anything inside the kernel_main function since an IDE checking
the code with tests enabled would be less than helpful due to the
function not being compiled in the current configuration.