Fix crash on interrupt before scheduler initialized
In a previous change, code was added to the end_interrupt function to
ensure that threads woken up by the interrupt would actually run instead
when the core was otherwise idle. However, this inadvertently broke if
an interrupt occurred before the scheduler was fully initialized, as the
kernel process would not be allocated at that point.

To avoid this, a check has been added to skip the task switch when the
scheduler has not yet been initialized.