Make DeviceRef and DeviceWeak wrapper structs
Previously, DeviceRef and DeviceWeak were just defined as type aliases
for Arc/Weak. This exposed some implementation details unnecessarily and
required that [DeviceNode<$>] be specified whenever using dyn_dyn_cast!.

By changing these into actual wrapper structs, we can choose what parts
of Arc/Weak to actually expose and avoid the need to ever specify that
DeviceNode is present when using dyn_dyn_cast!.

Additionally, this has the added benefit of allowing a custom Debug
implementation to be provided. We now do this so that we print the
device name instead of trying to actually dereference and print the
device (which can lead to infinite recursion in devices that have
circular references).