Change up the get_phys_mem_ptr API
Previously, this API returned a simple pointer. However, it's possible
that an implementation for a future architecture might need to map
physical memory only as needed instead of mapping all physical memory at
boot. To ensure this remains possible without massive API changes, this
function now returns a PhysMemPtr type which is permitted to be used to
manage such mappings.