Use soft interrupts for logging
Previously, messages logged using log! would be logged immediately. This
could cause problems with trying to recursively lock spinlocks if
anything dealing with TTYs that were used for logging was itself trying
to log. While it can cause endless logging in some cases, it's actually
perfectly fine in some, e.g. if only the code for reading input is
logging, so allow this to work by using a soft interrupt to delay
logging.