Fix soft interrupts happening in reverse order
Due to the fact that a stack was previously being used to store soft
interrupt functions, whenever soft interrupts were run, they would be
run in the reverse of the order they were enqueued in. While most soft
interrupts don't really care what order they're run in, this means that
messages logged via log! would not appear in the correct order if logged
when an InterruptDisabler was held.