Move GlobalAlloc machinery out of early alloc
Previously, the machinery for the global allocator was in the early
allocation module. This made sense at the time, since there was not yet
proper memory allocation.

In preparation for adding proper memory allocators that actually use the
underlying frame and virtual memory allocators, move the machinery for
the global allocator into the parent `mem` module and set it up to allow
the allocator to be switched later.