Add a way of setting kernel options
A new framework has been added to be able to set runtime options
controlling how the kernel should behave. Currently, it uses env! to
include the value of an environment variable at compile time, but it
should be possible to relatively easily switch to using an options
string from the bootloader if/when we ever move to a bootloader that can
support such a thing.