Try harder to keep testing after failed test
Previously, testing would always terminate if a test panicked while an
InterruptDisabler was held. While this was technically the safest option
(it's never possible to prove that it's safe to continue since the panic
might occur in a location when critical data structures are locked or
otherwise corrupted), most of the time such InterruptDisablers come from
the test needing to hold a spinlock on one of its own data structures
that it constructed and that is not critical to system operation.

To allow continued testing on a best effort basis, just forcibly drop
any InterruptDisabler if a test panics. This is somewhat unsafe, but
this only occurs in a test environment and a lot of test failures could
indicate that the kernel would exhibit unsafe behaviour in general
anyway.

For data structures locked by an UninterruptibleSpinlock which were
held, we can rely on spinlock tracking to prevent the tests from
deadlocking anyway, since the spinlocks are not released.