Add a Makefile for high-level builds
While I was hoping to be able to avoid this for as long as possible and
use Cargo as much as possible, it turns out that Cargo's support for
some of the features I'll need in the future when building leaves quite
a bit to be desired.

For instance, getting binaries that need to be compiled for different
targets to co-exist in the same workspace seems to be extremely
difficult if not impossible while also being able to use build-std. This
will make it difficult in the future once we want to be able to compile
binaries for user-space (which would almost certainly want to use a
different target triple from kernel code), as well as once we want to
create an initialization ramdisk of some kind (which would need to be
packed on the host).

To allow all of this, just add a top-level Makefile that will be used to
invoke Cargo on subprojects and deal with some top-level tasks that will
need to be done on the host machine during a full system build.