Add separate grow and shrink methods to FixedBitVector
Previously, we had a single FixedBitVector::resize method. This could be
used to either grow or shrink the BV, but it could make it difficult to
infer the intent of some code, since growing and shrinking are really
slightly different operations even if they're both just "resizing" the
underlying BV.

To address this, add new grow and shrink methods that have where clauses
ensuring the relationship between the sizes of the old and new sizes.