Change SlabAlloc to use a type instead of a size
Previously, SlabAlloc took the size of the objects to allocate instead
of a type to actually allocate. This required an extra, annoying where
clause that ended up polluting a lot of code and just generally made it
more difficult than necessary to work with SlabAlloc directly.

Instead, just make SlabAlloc take a type T that it should allocate
memory for. Users can still provide [u8; SIZE] to get the old behaviour
anyway.