Add a list of all active slab allocators
Previously, hkd would have to manually be updated with every individual
slab allocator that was added. Instead, add a global list of active slab
allocators that the command line can use. This could also be useful in
the future for other purposes, e.g. we could try to free empty slabs
when memory pressure is high.