Fix SlabInfo free BV setting too many free bits
Before a previous change that simplified how FixedBitVector's new method
works, the SlabInfo::new method would only set the first n bits as free
since no objects exist beyond that in the slab. However, in an attempt
to clean things up, this was inadvertenly changed so that all bits would
be set regardless of n.

This change adds a set_range method to FixedBitVector and uses that
instead to ensure that only the first n bits are set.