Fix incorrect num_free on SLAB_INFO
When marking the space that will be used by the SlabInfo struct as not
free inside slab allocators with OWN_INFO set, the num_free value was
inadvertently not being decremented. This could result in incorrect
behaviour, since the allocator assumes that num_free accurately reflects
the number of free objects in the slab.

In combination with the just-fixed issue where too many bits were set in
the free BV, this could result in an object being allocated after the
end of the slab and cause memory corruption.