Fall back to early allocator when panicking
Previously, the code for dealing with panics would use whichever global
allocator was currently in use. This was fine most of the time, unless
the crash happened inside the memory allocator itself. While the current
panic handler does not allocate any memory, it's not hard to imagine
that more sophisticated handling (which would be useful for debugging)
may need to allocate small amounts of scratch memory and doing so may
deadlock or trigger a recursive panic if the memory allocator itself
panicked.

Instead, we'll fall back to the early memory allocator, whose region
should still be available. This allocator should be much safer to use,
even in the event of a very unstable system.